log: init cube
A cube has been rendered. This commit contains the code for drawing a
cube onto the canvas. The cube is with a function createCube. Currently
only one cube can be drawn at a time since the EBO directly uses the
faces of the cube to render said cube, introducing more cubes would
cause an overlap of face indexes and they will reference the incorrect
vertex within the VBO.

A likely solution would be to provide a VBO and EBO per "mesh" or cube
in this case, and iterating through each EBO in the while loop, but that may not be the most efficient way of doing things.

Next: Camera Movement!