log: add camera rotation w/ euler angles
The camera can now be tilted up and down (pitched) and moved left and
right (yawed) as opposed to just walking like before (wasd movement). I
suck at math now and picturing how these things work in my head without
drawing them out and even then I struggle to understand what is going
on, so most of the code is in large part thanks to:
https://learnopengl.com/Getting-started/Camera

I would like to change the camera movement so that the cursor is locked
to the center of the screen like an FPS allowing for rotation beyond
what is currently allowed as it is bound by the avail width/height of
the canvas.