log: add oribiting camera
In this commit I changed the camera to be an orbiting camera instead of
the fps like camera I had before. The rotation along the xz plane works
fine for the most part, though it can be improved. The rotation along
the yz plane however, is broken. Whenever the rotation reaches 90
degrees in either direction, the camera flips??? I'm not sure what is
going on, its very difficult for me to understand what is even
happening. What I do know is that this weird behavior is caused by the
fact that the up vector for the lookat and the direction vector become
parallel, which causes the lookAt calculation for computing the right
and also the up prime vectors to become <0,0,0>. I don't know how to fix
it, I tried for a bit but want to move on to other things. I'm a bit
tired of dealing with camera rotation and movement!