log: add ability to move object
I've added functions/event callbacks that give me the ability to click an object (in this case just a cube), and move it to a new location.
Due to some timing issues, I can't just click again to drop the item in the new place, instead I have to press escape. This will hopefully be fixed in the next commit.