log: show AABB as wireframe
I was able to get the AABB to show up as a green wireframe as I hope
to implement the ability to customize the size of the AABBs but there
is an issue:

Since the AABB is computed after applying the model matrix to the
local space AABB coordinates, in order to get the min and max
and compute the correctly sized world space AABB, there is issues when
the actual cube to which the AABB belongs to is transformed further.

Trying to update the world aabb vetices via the model matrix
(only applying a translation for now) results in the aabb wireframe
being offset by the existing distance, essentially doubling the
translation.

As im typing this I think I found a solution: Compute the world AABB
vertices used ONLY FOR RENDERING, as if they were at the center of the
world, similar to how a cube's vertices are defined. And then translate
those vertices!

Next:
- fix aabb wireframe issue
- clean up some code
- Write code for importing actual models (reuse code from ShidE
Graphics Engine)