log: fix aabb wireframe transformation issue
An entity's AABB is now positioned correctly. The solution was to treat
the actual AABB (the thing one might click on) and the AABB wireframe
as two different things.

The actual AABB should follow the cube and
thus be transformed accordingly. Because it doesn't go through the
vertex shader since it is not related to actual graphics, it doesn't
undergo the MVP transformation again.

The AABB wireframe on the other hand, is essentially another entity,
and will be rendered via processing through the vertex shader,
meaning a proper model matrix should be assigned to it. Thus, the vertices
used to define said wireframe should be positioned at the center of the
world and later transformed (translated in this case).

Since translation is the only entity transformation allowed right now
(clicking and moving an entity), its the only transformation supported.
Ideally, I can do the same for scaling and rotations, though doing
either of these transformations on the primitive will require
recomputation of the AABB min/max, and the following computations to
generate axis aligned vertices...

Next: move "data" between cubes (routers)