log: move "data" between two "routers"
I added the ability to select two entities to have a cube move
between the two. This is meant to simulate the appearance of "data"
moving two "routers" for the project I have in mind. For now the
implementation is really bad. For one, if an object is moved, the
direction, or path that the "data" should take is not updated. And it
also just looks bad. This was mainly just to see how it might be done
in the future.

I need to start considering how I am going to manage state. Currently
I am kind of juggling creating a game engine, and also creating the
"game" (or educational tool) that I had in mind. Because of this I
am moving quite quickly (assuming I don't spend 3-4 days trying to
understand a concept I am implementing), and not really mapping out what
I need to do for the GAME ENGINE vs the GAME.

Next: improve code???