log: improve camera pan
The panning of the camera was improved to support various camera
locations. Previously the camera was moved by changing the x and y
coordinates of the camera. Since its initial position is at [0, 3, 10]
this worked as expected but changing the inital position to [10, 3, 0]
caused the camera to be panned along the x axis instead of the z axis as
it should.