log: fix ray origin
After "completing" camera orbiting and camera panning I forgot to update
the origin from which the ray is "shot" from. This caused some errors
with the ray which have now been fixed. Regarding the camera orbiting, I
left a TODO reminding me to go back and fix it as it has some unexpected
behavior I'm not entirely sure how to fix. And upon hitting 90 degrees
on the pitch, there is a brief error in which the contents are flipped
likely due to how the CAM_RIGHT vector is calculated at that time.