log: add camera zoom
Added a camera zooming feature. There were two methods to zoom in.
1) Change the fov, by making the fov smaller or larger you can zoom in
   and out respectively.
2) Move the camera position "back" and "forth" using the camera's
   direction vector

I opted for the second option as it falls neatly with the panCamera and
orbitCamera function which both alter the camera position vector. Also
I already knew how I would do this method, whereas the FOV method would
require a bit more thinking on my end to ensure I implemented it
correctly. It would also require me to update the projection matrix
which I didn't want to do for the above reason of keeping camera changes
to only the view matrix.