log: add more objects
I added more objects to the refactored code to ensure that it still
works. Again, there are probably better ways to handle things like the
list of VAOs needed for the renderer and actually setting up to render
an object, but for now this works. Next I need to reintroduce the ray
picking.