log: add boundingbox class
I wanted to separate the logic for creating axis aligned bounding box
from the actual object itself. The reason for this was that later down
the line I will want to add a way to adjust bounding boxes, to the
model so I figured I'd need it to be its own entity instead of being
tied to a model. What I added is just a skeleton so that I can remember
what I did within the old Object class to generate the AABBs.