log: begin mesh/model import feature
I added an file input to allow me to add new meshes/models to the scene.
I had to learn a bit about type="file" for inputs and about the File and
FileReader object.

Next I need to create a new Object that has the newly imported mesh.
That includes the VAO, AABB, etc.