log: add gizmo interactability
The gizmo will now pop up when an object is selected. Furthermore
clicking within the gizmo's bounds (the circle) will allow you to move
the object similar to the translate gizmo in blender. However there are
many bugs with it. Many of these bugs are likely caused by the odd way
in which I handle user interaction. So that is what I'm moving my focus
to.