log: fix gizmo interaction with off centered obj
The issue mentioned in the previous commit has been
fixed. It turns out that a "mouseup" triggered by
mouse0 is the same as a "click" event. Knowing this
I had to move the if statement that sets
gizmo_interact = false and returns to the "click"
listener before all the other code is executed.