log: add arrow gizmo
I created a simple 3D arrow model and imported it.
It follows the center of the currently selected object.

I've also fixed an issue where moving the camera position
in any way caused the circular translate gizmo to be
out of place due to the new center not being calculated
in relation to the screen space. The fix required
calculating the new view matrix after camera movement
(pan, orbit, zoom) whereas before it was calculated
before each call to frame() (the render function).