log: improve translation gizmo
I'm actually so frustrated right now. The previous
method for translating an object along a single axis
worked but it didn't seem to have the effect I wanted.

I then realized that the effect I wanted is essentially
what the circle gizmo does but just restricted to the
desired axis. So essentially that entire fiasco
of coming up with the intersection of two new lines
and what not was a waste of time.

GG.