log: fix minor bug with the axis gizmo
There was a minor bug where upon using the axis
after a previous use, the object would teleport to
a different location along the active axis.

The reason this occured was because the position
along an axis is computed via the difference between
the coordinates when clicking on the axis gizmo and
the coordinates of the mouse in world space.

The fix was to record the last static position of the
object and add that to the difference.

Fix: prev_pos + new_pos - start_pos