log: begin work on axis rotation gizmos
I just spent this time importing the model and trying
to figure out what might be the best way to rotate
based on the mouse position.

For now I just rotate based on some delta along an
arbitrary axis but the plan is to calculate the
angle between two points on the screen, that being
my start position (when I begin clicking the torus)
and my current position (where my mouse is while
holding m0).