log: add y axis rotate gizmo functionality
The y axis rotate gizmo tool now works as intended
for the most part. However, its not very clear if the
objects are rotating at the desired rate. That may need
to be worked on. Since I'm working on my laptop and
can't use the scroll button I can't orbit the camera
around to test how the axis rotation gizmo works under
certain circumstances but for now I'll ignore that.