log: remove cur_selection_prev_x
Transforming the object with the axis tools required the previous STATIC
transform of the object. I did this by having a variable for each
transform saved whenever the object was altered in any way.

Now these values are stored within the object itself as an object with a
property for pos, rotate, and scale.