log: remove dependencies on aabb_vao
The AABB wireframe mesh now needs to be preloaded onto the renderer
object. This populates the aabb_mesh property with the mesh and creates
a VAO for the mesh for rendering.

This removes the need for passing the AABB_vao to the SceneObject
constructor like before

Issues Addressed
* Improve VAO management (#3)
No more dependency on AABB_vao. No longer have to provide as an argument
to the SceneObject and  AxisAlignedBoundingBox classes.