log: change rotation 3d gizmo functionality
I've decided I don't want to deal with the toruses that I was originally
using for the 3d rotate gizmos. Instead I will use an arrow like object
and calculate an angle based on the distance between the start_pos and
cur_pos along the chosen axis. This way all logic is similar for all
transformations.