log: gizmo changes (3d rotate and 2d scale)
The 3d rotate gizmos have been changed to be similar to the 3d gizmos
for translating and scaling. Instead of using toruses like Blender or
Unity I opted to use arrow-like gizmos to maintain simplicty.

The 2d scale tool has finally been fixed. Now the distance between the
center of the 2d gizmo and the edge of the circle (the radius) is
calculated in world space to normalize the distance between the center
and the current positon of the mouse. This way whenever the mouse is
over the 2d gizmo UI the scaling factor should be about 1 as expected
reducing the odd sudden jumps in scale upon frequent interactions of the
2d scale gizmo.