log: fix minor issue with 2D translate gizmo
There was a minor issue with the 2d translate gizmo where the object
being translated would snap to the cursor when the 2d gizmo was clicked
and moved. While not very apparent with the smaller circle, with larger
circles it was very obvious.

Fix:
The fix was to calculate the offset from the start_pos to cur_pos
(cur_pos - start_pos) which is what all other gizmos do essentially.