log: remove debug line and plane
The debug line and plane were removed. The code was tacked on and not
what I want in the main file for now. Hopefully in the future I find a
better way to implement a debug feature. It will help a lot with
debugging the 2D and 3D gizmo issues.