log: improve character controller
The player object can now be moved laterally. Originally the
last_static_transform property of a SceneObject was updated after all
movement inputs had been processed resulting in the last update (the
later if statements) to be the final influence on the object's movement
per frame.

The solution was to update the last_static_position in each if statement
so that subsequent if statements will add movement to the NEW value.