log: populate scene objects list UI
The UI now correctly displays the objects that are loaded into the
scene. I want to add the ability to "select" the object by clicking on
its corresponding list UI element but the "selection" aka displaying the
gizmos is more involved and is tempting me to just implement a Pub / Sub
system for state changes. However the last time I had to do this, things
got a bit too complicated for my liking.