log: progress on model preview
I created another test area to get a better understanding of the
viewport / scissor relationship to see why that might be working against
me for creating the model preview.

It helped me understand that I have to disable the SCISSOR_TEST before
rendering to the new framebuffer. I don't entirely understand why as I'm
not 100% what the SCISSOR_TEST really is, but it works now!