log: fix ViewWindow sizing problem
I implemented a quick fix for the  ViewWindow sizing problem where view1
would end up in the wrong spot because of the order in which the div's
are assigned their sized. This caused problems where the render would
not be in the expected location.

The fix was to assign the left and bottom property for a ViewWindow
AFTER both windows had their width and height assigned.