log: use Raycaster in TransformControls
I replaced the global variable raycaster that was just an object with a
direction and origin property with the Raycaster class object to unify
the code a bit more. There are still sections that I think can be
polished, especially the hoverColorChange() function since its still
using a direct call to isIntersecting which I don't like, but that
requires that I find a better way to do all of the color checks so I'm
leaving it as is for now.