log: improve TransformControls mouseHover()
The mouseHover() function in TransformControls has been improved to
dispatch less "change" events.

Previously a "change" event would be sent whenever the mouse was moved
around (hovered) because I was constantly checking for intersections
with the mouse and the transform handles.

Now the "change" event is only dispatched if the axis changes from the
previous axis. I'm not a huge fan of this method and should probably
just do what three.js does with the custom set() function in Object.defineProperty() to detect changes.

The only issue is that the 2D gizmo is causing me problems. I wan't to
completely overhaul that damn thing in the future anyway so for now I'll
leave it as is.