log: implement object / ui updates on render
Now SceneObjects and the 2D gizmo are updated for each render call. The
way I went about this wasn't the best but it works for now. I hope to
improve it, especially for the 2D gizmo as I just create a function
inside of the main_gizmo object. Ideally 2D elements have their own
class this way I support more than just the 2D transform gizmo for
future use!