log: change map in SceneHierarchy
The most notable change is that SceneHierarchy only has one map as
opposed to two. The map's key, value pairs are explained in the comment
next to it. Since I added IDs to objects I could forego some of the
weird things I did with object names.

Also added selection distinction via AABB's being shown on the selected
object. This causes some weird behavior with AABB property in that
trying to set the object's show_AABB property to true and then swapping
off of the object (deselecting) will set it to false.