log: improve mesh assignment in SceneObjects
I had mentioned that, due to the hasty implementation of fromJSON and
toJSON, model importing, models were queried by the name of the
SceneObject. This meant that renaming the object would result in no mesh
data being retrieved since a model with that new name probably doesn't exist.

The fix was to replace SceneObject.mesh = {vertex info...} to
SceneObject.mesh = {name: "example", data: {vertex info...}} this way
the name of the model is assigned and used when using toJSON(). I didn't
actually test renaming things so lets do that now!