log: implement clear() and loadScene()
Not much to say here. The functions are self explanatory and just add
more completeness to the editor. I will say that it would be beneficial
to deactive any signals that are dispatched en masse similar to how I do
for loading scenes (adding models and objects) and clearing the editor
(removing models and objects).