log: work on lighting
I attempted to add simple Phong lighting to the editor
so that objects with a single color are more
distinguishable. I got it working on the testbench
but there seem to be issues with implementing it in
the actual editor / game engine. I will need to figure
these out later.