log: fail to use lighting
I have no clue what is going on. Something in the new
implementation is causing things to break, but just with
vertex colors. The lighting basically worked, except
it looked a bit funky but its probably caused by my use
of the normal matrix, which I can probably omit for now
for testing purposes. However, the issue was that the
vertex colors just weren't being assigned to the object.
If I manually set a vec3 value for the color, the render
would reflect that, but if I tried to use the vertex
color it would just make the object black. Very curious.

I at least got the lighting to work WITH vertex colors
in the testbench, so thats nice.