log: improve ply parser 2
I've changed parsePLY() from the PlyFile class to parse
all the properties that can be on a vertex regardless
of whether the file has them or not. This includes:
* vertex positions
* vertex colors
* vertex normals
* vertex UV coords

If the file does not have one of the properties then it
added to the mesh data as "null" this way I can use this
in the renderer to determine which VBOs to create for
that object to "dynamically" create VAOs.