log: rework Renderer shaders / programs
I realized I'm going to need multiple programs to be
able to handle different meshes. Some meshes may have
normals while others don't, which is fine but I want
things to continue working as is with no need for manual
inputs to get things to work. So VAOs need to be created
dynamically to accomodate whatever data the mesh has,
and the programs need to be used dynamically as well.

Right now I'm just doing an overhaul of how I store
programs and program variables as its a bit of a mess
at the moment.