log: mistake in last commit
There was a minor mistake in the previous commit. I
said that the transform controls were not being
rendered despite no errors showing up. I think I just
forgot to uncomment the part of the code responsible
for rendering them.

I've also added back the AABB outline.

The framerate issue still persists. Its awful on
my laptop. Further testing on my computer is required.

One definitive reason for the bad framerate is due
to the way the AABB is updated whenever the object's
transform is changed. The model matrix is applied to
each vertex of the model. The problem is that the
framerate is very bad when just moving the camera
which should not update the model matrix and therefore
not do the expensive AABB recalculations.