log: complete Renderer rework
Renderer has now been updated to support meshes with normals. I still
feel like I use a lot of hacky code, and as mentioned in previous
commits, the performance (on my laptop) is awful. On my computer it
works just fine but this thing is pretty beefy, also my laptop's battery
imploded or something and I can't even run any games on it so that could
be why.

To test I would need to log framerates under the current implementation
and do the same for the previous implementation that didn't support
meshes with normals.