log: work on textures
Just testing out textures on the testbench. I need to flip the image
data and learn more about what some of the GL texParameteri() arguments
mean as well as get a better understanding of binding textures and
passing them to the fragment shader via the uniforms.

Additionally, the lack of seams in both the texture and lighting tests
is concerning since the seams are very noticeable in the actual editor.
They only seem to appear on objects with normals and are visible on BOTH
objects using their vertex colors and objects using a solid set color.