log: add texture / image importing
Textures (or images) can now be imported. The code to support this
feature was super rushed, im surprised it even works. Currently the
images are flipped upside down since they share the same classname as
model cards which have the model's image taken upside down, then flipped
via css property: transform: scaleY(-1). In the future I need to create
a function for flipping image data since the textures aren't mapped to
an object correctly for the same reasons. An example of this can be seen
in testbench/textures.